+++ src/test-embedding/java/org/truffleruby/test/embedding/ContextPermissionsTest.java
@@ -86,13 +86,13 @@ public class ContextPermissionsTest {
 
     @Test
     public void testThreadsNoNative() throws Throwable {
-        // The ruby.single_threaded option needs to be set because --single-threaded defaults to --embedded.
         try (Context context = Context
                 .newBuilder("ruby")
                 .allowCreateThread(true)
-                .allowExperimentalOptions(true)
-                .option("ruby.single-threaded", "false")
                 .build()) {
+            String option = "Truffle::Boot.get_option('single-threaded')";
+            Assert.assertEquals(false, context.eval("ruby", option).asBoolean());
+
             Assert.assertEquals(3, context.eval("ruby", "1 + 2").asInt());
 
             Assert.assertEquals(7, context.eval("ruby", "Thread.new { 3 + 4 }.value").asInt());
@@ -111,8 +111,6 @@ public class ContextPermissionsTest {
     public void testNoThreadsEnforcesSingleThreadedOption() throws Throwable {
         try (Context context = Context
                 .newBuilder("ruby")
-                .allowExperimentalOptions(true)
-                .option("ruby.single-threaded", "false")
                 .build()) {
             String option = "Truffle::Boot.get_option('single-threaded')";
             Assert.assertEquals(true, context.eval("ruby", option).asBoolean());